test: rename dir and file
